reconcile to a single demo project with simple docs pages
