+++ docs/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@mdx-js/react": "^1.6.1",
     "formidable-oss-badges": "^0.3.5",
+    "marksy": "^8.0.0",
     "prism-react-renderer": "^1.1.1",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
@@ -22,10 +23,8 @@
     "react-router-dom": "^5.2.0",
     "react-router-ga": "^1.2.3",
     "react-static": "^7.2.3",
-    "react-static-plugin-md-pages": "^0.1.7",
     "react-static-plugin-react-router": "^7.2.3",
-    "react-static-plugin-sitemap": "^7.2.3",
-    "react-static-plugin-source-filesystem": "^7.4.2",
+    "react-static-plugin-sitemap": "^7.0.0",
     "react-static-plugin-styled-components": "^7.2.2",
     "styled-components": "^5.1.0"
   },
