--- helpers-demo/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["react-static/babel-preset.js"]
-}
