--- helpers-demo/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "helpers-demo",
-  "private": true,
-  "scripts": {
-    "start": "react-static start",
-    "stage": "react-static build --staging",
-    "build": "react-static build",
-    "analyze": "react-static build --analyze",
-    "serve": "serve dist -p 3000"
-  },
-  "dependencies": {
-    "@reach/router": "^1.2.1",
-    "axios": "^0.19.0",
-    "marksy": "^8.0.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-static": "^7.4.2",
-    "react-static-plugin-reach-router": "^7.4.2",
-    "react-static-plugin-sitemap": "^7.4.2",
-    "react-static-plugin-source-filesystem": "^7.4.2",
-    "slugify": "^1.4.5"
-  },
-  "devDependencies": {
-    "serve": "^11.1.0"
-  }
-}
