--- helpers-demo/src/App.js
@@ -1,33 +1,0 @@
-import React from "react";
-import { Root, Routes, addPrefetchExcludes } from "react-static";
-//
-import { Link, Router } from "components/Router";
-import Dynamic from "containers/Dynamic";
-
-import "./app.css";
-
-// Any routes that start with 'dynamic' will be treated as non-static routes
-addPrefetchExcludes(["dynamic"]);
-
-function App() {
-  return (
-    <Root>
-      <nav>
-        <Link to="/">Home</Link>
-        <Link to="/about">About</Link>
-        <Link to="/docs">Docs</Link>
-        <Link to="/dynamic">Dynamic</Link>
-      </nav>
-      <div className="content">
-        <React.Suspense fallback={<em>Loading...</em>}>
-          <Router>
-            <Dynamic path="dynamic" />
-            <Routes path="*" />
-          </Router>
-        </React.Suspense>
-      </div>
-    </Root>
-  );
-}
-
-export default App;
