--- helpers-demo/src/components/Router.js
@@ -1,1 +1,0 @@
-export { Link, Router } from "@reach/router";
