--- helpers-demo/src/containers/Dynamic.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function Dynamic() {
-  return (
-    <div>
-      This is a dynamic page! It will not be statically exported, but is
-      available at runtime
-    </div>
-  );
-}
