--- helpers-demo/src/pages/404.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function _404() {
-  return (
-    <div>
-      <h1>404 - Oh no! We could not find that page :(</h1>
-    </div>
-  );
-}
