--- helpers-demo/src/pages/about.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function About() {
-  return (
-    <div>
-      <p>React Static is a progressive static site generator for React.</p>
-    </div>
-  );
-}
