--- helpers-demo/src/pages/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return (
-    <div style={{ textAlign: "center" }}>
-      <h1>Welcome to React-Static</h1>
-    </div>
-  );
-}
